Move splitter to own type
